Initclose.h kendu da, programa txukundu da
