+++ MUwar/MUwar.vcxproj.filters
@@ -24,9 +24,6 @@
     <ClCompile Include="src\mapa.c">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="InitClose.c">
-      <Filter>Source Files</Filter>
-    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\tropa.h">
@@ -36,9 +33,6 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="src\mapa.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="InitClose.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
